@@ -19,20 +19,21 @@
   },
   "homepage": "https://github.com/shuding/nextra-docs-template#readme",
   "dependencies": {
-    "autoprefixer": "^10.4.15",
+    "autoprefixer": "^10.4.20",
     "flexsearch": "0.7.31",
-    "next": "^13.4.19",
+    "next": "^15.1.3",
     "next-sitemap": "^4.2.3",
-    "nextra": "^2.12.3",
-    "nextra-theme-docs": "^2.12.3",
-    "postcss": "^8.4.29",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "tailwindcss": "^3.3.3"
+    "nextra": "^3.3.1",
+    "nextra-theme-docs": "^3.3.1",
+    "pnpm": "^9.15.2",
+    "postcss": "^8.4.49",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "tailwindcss": "^3.4.17"
   },
   "devDependencies": {
     "@types/node": "20.6.0",
-    "typescript": "^5.2.2"
+    "typescript": "^5.7.2"
   },
   "pnpm": {
     "overrides": {
