@@ -1,14 +1,22 @@
-const withNextra = require('nextra')({
+import nextra from 'nextra'
+
+const withNextra = nextra({
   theme: 'nextra-theme-docs',
   themeConfig: './theme.config.tsx',
   latex: true,
+  search: {
+    codeblocks: false
+  },
   defaultShowCopyCode: true
 })
 
-module.exports = withNextra({
+
+
+export default withNextra({
 	trailingSlash: false,
 	basePath: '/docs',
-	assetPrefix: '/docs',
+  assetPrefix: '/docs',
+  reactStrictMode: true,
   async redirects() {
     return [
       {
