@@ -1,6 +1,7 @@
 /** @type {import('next-sitemap').IConfig} */
 module.exports = {
     siteUrl: 'https://memgraph.com/docs',
-    generateIndexSitemap: false
-  }
+    generateIndexSitemap: false,
+    exclude: ['*/_meta']
+}
   
