@@ -1,7 +1,7 @@
-{
+export default {
   "available-algorithms": "Available algorithms",
   "deep-path-traversal": "Deep path traversal algorithms",
   "install-mage": "Install MAGE",
   "run-algorithms": "Run algorithms",
   "utilize-networkx": "Utilize NetworkX"
-}
+};
