@@ -1,4 +1,4 @@
-{
+export default {
   "algo": "algo",
   "betweenness_centrality_online": "betweenness_centrality_online",
   "betweenness_centrality": "betweenness_centrality",
@@ -61,4 +61,4 @@
   "vrp": "vrp",
   "weakly_connected_components": "weakly_connected_components", 
   "xml_module": "xml_module"
-}
+};
