@@ -1,4 +1,4 @@
-{
+export default {
     "graph-rag": "GraphRAG",
     "machine-learning": "Machine learning"
 }
