@@ -1,4 +1,4 @@
-{
+export default {
     "high-availability": "High availability",
     "replication": "Replication"
 }
