@@ -1,4 +1,4 @@
-{
+export default {
   "c-api": "C API",
   "c-example": "C example"
 }
