@@ -1,4 +1,4 @@
-{
+export default {
   "cpp-api": "C++ API",
   "cpp-example": "C++ example"
 }
