@@ -1,4 +1,4 @@
-{
+export default {
   "python-api": "Python API",
   "mock-python-api": "Mock Python API",
   "python-example": "Python example",
