@@ -1,4 +1,4 @@
-{
+export default {
   "rust-api": "Rust API",
   "rust-example": "Rust example"
 }
