@@ -1,4 +1,4 @@
-{
+export default {
   "best-practices": "Best practices",
   "csv": "CSV",
   "json": "JSON",
