@@ -1,4 +1,4 @@
-{
+export default {
   "transformation-modules": "Transformation modules",
   "manage-streams-query": "Manage streams with queries",
   "manage-streams-lab": "Manage streams with UI",
