@@ -1,4 +1,4 @@
-{
+export default {
   "built-in-elements": "Built-in elements",
   "directive-properties": "Directive properties"
 }
