@@ -1,4 +1,4 @@
-{
+export default {
   "authentication-and-authorization": "Authentication and authorization",
   "backup-and-restore": "Backup and restore",
   "configuration": "Configuration",
