@@ -1,6 +1,6 @@
-{
+export default {
     "users": "Users",
     "role-based-access-control": "Role-based access control",
     "auth-system-integrations": "Auth system integrations"
-  }
+}
   
