@@ -1,4 +1,4 @@
-{
+export default {
     "docker": "Docker",
     "linux": "Linux",
     "aws": "AWS",
