@@ -1,4 +1,4 @@
-{
+export default {
   "constraints": "Constraints",
   "data-types": "Data types",
   "data-durability": "Data durability",
