@@ -1,4 +1,4 @@
-{
+export default {
   "install-memgraph": "Install Memgraph",
   "cli": "CLI",
   "first-steps-with-docker": "First steps with Docker"
