@@ -1,4 +1,4 @@
-{
+export default {
   "docker": "Docker",
   "debian": "Debian",
   "ubuntu": "Ubuntu",
