@@ -1,4 +1,4 @@
-{
+export default {
   "faq": "FAQ",
   "errors": "Errors"
 }
