@@ -1,4 +1,4 @@
-{
+export default {
   "alter": "ALTER",
   "call": "CALL",
   "case": "CASE",
