@@ -1,4 +1,4 @@
-{
+export default {
   "analyzing-ted-talks": "Analyzing TED talks",
   "backpacking-through-europe": "Backpacking through Europe",
   "exploring-the-european-road-network": "Exploring the European road network",
